refactor: move holder logic from operations.go
closes #428
