+++ pkg/restapi/internal/common/http/common.go
@@ -1,0 +1,41 @@
+/*
+Copyright SecureKey Technologies Inc. All Rights Reserved.
+
+SPDX-License-Identifier: Apache-2.0
+*/
+
+package http
+
+import (
+	"encoding/json"
+	"io"
+	"net/http"
+
+	log "github.com/sirupsen/logrus"
+)
+
+// ErrorResponse to send error message in the response
+type ErrorResponse struct {
+	Message string `json:"errMessage,omitempty"`
+}
+
+// WriteErrorResponse write error resp
+func WriteErrorResponse(rw http.ResponseWriter, status int, msg string) {
+	rw.WriteHeader(status)
+
+	err := json.NewEncoder(rw).Encode(ErrorResponse{
+		Message: msg,
+	})
+
+	if err != nil {
+		log.Errorf("Unable to send error message, %s", err)
+	}
+}
+
+// WriteResponse writes interface value to response
+func WriteResponse(rw io.Writer, v interface{}) {
+	err := json.NewEncoder(rw).Encode(v)
+	if err != nil {
+		log.Errorf("Unable to send error response, %s", err)
+	}
+}
