+++ pkg/restapi/vc/operation/models.go
@@ -11,6 +11,8 @@ import (
 	"time"
 
 	"github.com/hyperledger/aries-framework-go/pkg/doc/verifiable"
+
+	"github.com/trustbloc/edge-service/pkg/restapi/model"
 )
 
 // CreateCredentialRequest input data for edge service issuer rest api
@@ -45,15 +47,9 @@ type ProfileRequest struct {
 	DIDPrivateKey           string                             `json:"didPrivateKey"`
 	DIDKeyType              string                             `json:"didKeyType"`
 	DIDKeyID                string                             `json:"didKeyID"`
-	UNIRegistrar            UNIRegistrar                       `json:"uniRegistrar,omitempty"`
+	UNIRegistrar            model.UNIRegistrar                 `json:"uniRegistrar,omitempty"`
 	DisableVCStatus         bool                               `json:"disableVCStatus"`
 	OverwriteIssuer         bool                               `json:"overwriteIssuer,omitempty"`
-}
-
-// UNIRegistrar uni-registrar
-type UNIRegistrar struct {
-	DriverURL string            `json:"driverURL,omitempty"`
-	Options   map[string]string `json:"options,omitempty"`
 }
 
 // VerifyCredentialResponse describes verify credential response
@@ -167,38 +163,4 @@ type VerifyPresentationCheckResult struct {
 	Check              string `json:"check,omitempty"`
 	Error              string `json:"error,omitempty"`
 	VerificationMethod string `json:"verificationMethod,omitempty"`
-}
-
-// ErrorResponse to send error message in the response
-type ErrorResponse struct {
-	Message string `json:"errMessage,omitempty"`
-}
-
-// HolderProfileRequest holder mode profile request
-type HolderProfileRequest struct {
-	Name                    string                             `json:"name"`
-	SignatureType           string                             `json:"signatureType"`
-	SignatureRepresentation verifiable.SignatureRepresentation `json:"signatureRepresentation"`
-	DID                     string                             `json:"did"`
-	DIDPrivateKey           string                             `json:"didPrivateKey"`
-	DIDKeyType              string                             `json:"didKeyType"`
-	DIDKeyID                string                             `json:"didKeyID"`
-	UNIRegistrar            UNIRegistrar                       `json:"uniRegistrar,omitempty"`
-	OverwriteHolder         bool                               `json:"overwriteHolder,omitempty"`
-}
-
-// SignPresentationRequest request for signing a presentation.
-type SignPresentationRequest struct {
-	Presentation json.RawMessage          `json:"presentation,omitempty"`
-	Opts         *SignPresentationOptions `json:"options,omitempty"`
-}
-
-// SignPresentationOptions options for signing a presentation.
-type SignPresentationOptions struct {
-	VerificationMethod string     `json:"verificationMethod,omitempty"`
-	AssertionMethod    string     `json:"assertionMethod,omitempty"`
-	ProofPurpose       string     `json:"proofPurpose,omitempty"`
-	Created            *time.Time `json:"created,omitempty"`
-	Challenge          string     `json:"challenge,omitempty"`
-	Domain             string     `json:"domain,omitempty"`
 }
